+++ pkg/openers/crng_test.go
@@ -1,11 +1,14 @@
 //go:build !fileonly
 // +build !fileonly
 
-package openers //nolint
+package openers_test
 
 import (
 	"io"
 	"testing"
+
+	"github.com/batmac/ccat/pkg/openers"
+	"github.com/batmac/ccat/pkg/utils"
 )
 
 func Test_crngOpener_Open(t *testing.T) {
@@ -22,8 +25,7 @@ func Test_crngOpener_Open(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			f := crngOpener{}
-			got, err := f.Open(tt.args, false)
+			got, err := openers.Open(tt.args, false)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("crngOpener.Open() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -46,40 +48,7 @@ func Test_crngOpener_Open(t *testing.T) {
 				return
 			}
 
-			checkRandomness(t, data)
-		})
-	}
-}
-
-func checkRandomness(t *testing.T, data []byte) {
-	t.Helper()
-	acc := uint8(0)
-	for i := 0; i < len(data); i++ {
-		acc |= data[i]
-	}
-	if acc != 0xFF {
-		t.Errorf("crngOpener.Open() = %v, want %v", acc, 255)
-		return
-	}
-}
-
-func Test_crngOpener_Evaluate(t *testing.T) {
-	tests := []struct {
-		name string
-		args string
-		want float32
-	}{
-		{name: "ko", args: "false://", want: 0},
-		{name: "ok", args: "crng://", want: 0.9},
-		{name: "ok with good limit", args: "crng://10", want: 0.9},
-		{name: "ok with good limit 2", args: "crng://10k", want: 0.9},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := crngOpener{}
-			if got := f.Evaluate(tt.args); got != tt.want {
-				t.Errorf("crngOpener.Evaluate() = %v, want %v", got, tt.want)
-			}
+			utils.CheckBytesRandomness(t, data)
 		})
 	}
 }
