+++ pkg/openers/prng_test.go
@@ -1,11 +1,14 @@
 //go:build !fileonly
 // +build !fileonly
 
-package openers //nolint
+package openers_test
 
 import (
 	"io"
 	"testing"
+
+	"github.com/batmac/ccat/pkg/openers"
+	"github.com/batmac/ccat/pkg/utils"
 )
 
 func Test_prngOpener_Open(t *testing.T) {
@@ -21,8 +24,7 @@ func Test_prngOpener_Open(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			f := prngOpener{}
-			got, err := f.Open(tt.args, false)
+			got, err := openers.Open(tt.args, false)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("prngOpener.Open() %s: error = %v, wantErr %v", tt.name, err, tt.wantErr)
 				return
@@ -46,27 +48,7 @@ func Test_prngOpener_Open(t *testing.T) {
 				return
 			}
 
-			checkRandomness(t, data)
-		})
-	}
-}
-
-func Test_prngOpener_Evaluate(t *testing.T) {
-	tests := []struct {
-		name string
-		args string
-		want float32
-	}{
-		{name: "ko", args: "false://", want: 0},
-		{name: "ok", args: "prng://", want: 0.9},
-		{name: "ok with good limit", args: "prng://10", want: 0.9},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := prngOpener{}
-			if got := f.Evaluate(tt.args); got != tt.want {
-				t.Errorf("prngOpener.Evaluate() = %v, want %v", got, tt.want)
-			}
+			utils.CheckBytesRandomness(t, data)
 		})
 	}
 }
